+++ packages/config/src/Config.ts
@@ -572,13 +572,12 @@ export function getDefaultTarget(projectRoot: string): ProjectTarget {
   if (exp.sdkVersion && exp.sdkVersion !== 'UNVERSIONED' && semver.lt(exp.sdkVersion, '37.0.0')) {
     return 'managed';
   }
-  return isBareWorkflowProject(projectRoot) ? 'bare' : 'managed';
+  return isBareWorkflowProject(projectRoot, exp) ? 'bare' : 'managed';
 }
 
-function isBareWorkflowProject(projectRoot: string): boolean {
-  const { pkg } = getConfig(projectRoot, {
-    skipSDKVersionRequirement: true,
-  });
+function isBareWorkflowProject(projectRoot: string, exp: ExpoConfig): boolean {
+  const [pkg] = getPackageJsonAndPath(projectRoot, exp);
+
   if (pkg.dependencies && pkg.dependencies.expokit) {
     return false;
   }